Add display of coffee shop data
