+++ views/coffeeShop.handlebars
@@ -4,11 +4,11 @@
 
 <main>
     {{#each coffeeShops}}
-        <article>
+        <article class="coffeeShopDetail">
             <strong>{{this.tags.name}}</strong>
             <p>{{this.tags.addr:housenumber}} {{this.tags.addr:street}}, {{this.tags.addr:city}}{{#if this.tags.addr:postcode}},
             {{this.tags.addr:postcode}}{{/if}}</p>
-            <p>Latitude: {{this.lat}}\tLongitude: {{this.lon}}</p>
+            <p>Latitude: {{this.lat}} | Longitude: {{this.lon}}</p>
         </article>
     {{/each}}
 </main>
